requirements/python3.10.txt
urllib3
pydantic==1.10.12
fastapi==0.95.0
fastapi-jwt-auth
PyYAML~=6.0
Jinja2~=3.1.2
watchdog~=2.1.8
humanize~=4.1.0
openxmllib-py3~=2.0.0
Cython~=3.0a1
wfastcgi~=3.0.0.0
a2wsgi~=1.4.1
python-jose~=3.3.0
passlib~=1.7.4
python-multipart~=0.0.5
pdfminer.six
pdfplumber~=0.7.1
bcrypt
kink~=0.6.5
hypercorn[trio]
gunicorn~=20.1.0
uvicorn[standard]
pika~=1.3.1
pythonnet
tika==1.25
py_vncorenlp
toolz==0.12.0
pandas
doctr==1.9.0
ocrmypdf==13.4.7
#easyocr==1.6.2
img2pdf==0.4.4
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.7.6
pdfreader==0.1.12
pika==1.3.1
pikepdf==7.1.1
Pillow==9.4.0
pillow-avif-plugin==1.2.2
textblob==0.17.1
underthesea
opencv-python
PyMuPDF
PyPDF2
moviepy
matplotlib
tika==1.25
numpy
scipy
torch --extra-index-url https://download.pytorch.org/whl/cpu
torchvision --extra-index-url https://download.pytorch.org/whl/cpu
torchaudio --extra-index-url https://download.pytorch.org/whl/cpu
git+https://github.com/facebookresearch/detectron2.git@2c6c380f94a27bd8455a39506c9105f652b9f760
gradio
easyocr==1.6.2
pillow-avif-plugin
underthesea
langdetect
keras-ocr
tensorflow
#fitz
#sudo apt install libpython3.10-dev
#python -m pip install -r requirements/python3.10.txt
